Canonical Host Check; Using HAVE_LIBCURLDLL Macro
Checks for host for which package is being built. Win32 hosts (MinGW) will check for libcurldll. All other POSIX hosts will check for libcurl.

In connection with checking for libcurldll on MinGW hosts, assurance is needed to determine not only if cURL headers are available, but if either libcurl or libcurldll is present as well.