Optional Feature Modified; Checking for Uppercase <SCRIPT> Tag; Modified generateLists() Warning
Replaced --enable-curl with --with-curl as the former pertains to the use of an optional feature within the code whereas the latter pertains to the use of an optional package external of the software.

Added charToLowerCase() in case stripMarkupTags() finds <SCRIPT> tags in uppercase.

Explicitly warns, whenever generateLists() is called, that if a given file is remote, to install libcurl.