Dynamic Allocation of Unsorted Words; Minor Modification of Search Display; Modified Uppercase SCRIPT Tag Checking
Tokens from file buffer were stored to list of unsorted words on the stack, causing buffer overflow when accessed repeatedly. Tokens are now passed to strdup to be allocated on the heap.

Prints theValue only when valueFlag and searchFlag are false (0).

Used charToLowerCase() incorrectly on literal characters. Should be used on characters of inputted string.