New Application for Interactivity
Created "eng-qbl," a new command line application that offers working with the English Kabbalah system interactively.

New function, listToString, takes a list of words from a struct wordlist and returns that list as a string of words delimited by spaces. This eliminates the tedious need to constantly use quotes when inputting phrases on the command line.